requirements.txt
# Core API dependencies
fastapi[all]
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
python-dotenv>=1.0.0

# Face recognition and image processing
face-recognition>=1.3.0
dlib>=19.24.0
numpy>=1.24.0
Pillow>=10.1.0